use next.js subpath i18n feature
